[workloadmeta] Extract completeness tracking into its own type
