+++ comp/core/workloadmeta/impl/cached_entity.go
@@ -8,6 +8,7 @@ package workloadmetaimpl
 import (
 	"maps"
 	"reflect"
+	"slices"
 	"sort"
 
 	wmdef "github.com/DataDog/datadog-agent/comp/core/workloadmeta/def"
@@ -111,4 +112,8 @@ func (e *cachedEntity) copy() *cachedEntity {
 	maps.Copy(newEntity.sources, e.sources)
 
 	return newEntity
+}
+
+func (e *cachedEntity) reportedSources() []wmdef.Source {
+	return slices.Collect(maps.Keys(e.sources))
 }
