+++ comp/core/workloadmeta/impl/store.go
@@ -165,7 +165,7 @@ func (w *workloadmeta) Subscribe(name string, priority wmdef.SubscriberPriority,
 					events = append(events, wmdef.Event{
 						Type:       wmdef.EventTypeSet,
 						Entity:     entity,
-						IsComplete: w.isEntityComplete(kind, cachedEntity),
+						IsComplete: w.completeness.isComplete(kind, cachedEntity.reportedSources()),
 					})
 				}
 			}
@@ -915,7 +915,7 @@ func (w *workloadmeta) handleEvents(evs []wmdef.CollectorEvent) {
 				filteredEvents[sub] = append(filteredEvents[sub], wmdef.Event{
 					Type:       wmdef.EventTypeSet,
 					Entity:     entity,
-					IsComplete: w.isEntityComplete(entityID.Kind, cachedEntity),
+					IsComplete: w.completeness.isComplete(entityID.Kind, cachedEntity.reportedSources()),
 				})
 			} else {
 				entity = entity.DeepCopy()
@@ -930,7 +930,7 @@ func (w *workloadmeta) handleEvents(evs []wmdef.CollectorEvent) {
 					Entity: entity,
 					// Same as with entity, completeness refers to the copy
 					// before the unset took place
-					IsComplete: w.isEntityComplete(entityID.Kind, cachedEntity),
+					IsComplete: w.completeness.isComplete(entityID.Kind, cachedEntity.reportedSources()),
 				})
 			}
 		}
@@ -1048,26 +1048,6 @@ func classifyByKindAndID(entities []wmdef.Entity) map[wmdef.Kind]map[string]wmde
 	}
 
 	return res
-}
-
-// isEntityComplete checks if an entity is complete, meaning all expected
-// collectors have reported data for it. If no expected sources are defined for
-// the entity kind, it returns true (considered complete by default).
-func (w *workloadmeta) isEntityComplete(kind wmdef.Kind, cachedEntity *cachedEntity) bool {
-	expectedSources, ok := w.expectedSources[kind]
-	if !ok || len(expectedSources) == 0 {
-		// No expected sources defined for this kind, consider it complete
-		return true
-	}
-
-	// Check if all expected sources have reported
-	for _, expectedSource := range expectedSources {
-		if _, reported := cachedEntity.sources[expectedSource]; !reported {
-			return false
-		}
-	}
-
-	return true
 }
 
 // resolveCollectorPullInterval returns the pull interval for a collector. If
